django/requirements.txt
certifi==2019.11.28
chardet==3.0.4
Django==2.2.6
django-cors-headers==3.2.1
djangorestframework==3.10.3
httpie==2.0.0
idna==2.8
mysql-connector-python==8.0.18
mysqlclient==1.4.6
Pillow==6.2.1
protobuf==3.10.0
Pygments==2.5.2
pytz==2019.3
requests==2.22.0
six==1.12.0
sqlparse==0.3.0
urllib3==1.25.7
